+++ TemplateGame/DrawDebug.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "define.h"
+#include "Texture.h"
 #include "DeviceManager.h"
 
 class DrawDebug
@@ -20,9 +21,9 @@ public:
 
 	void drawLine(Vec2, Vec2, DWORD = D3DCOLOR_XRGB(255, 0, 0), float = 2);
 
-	void drawLineHorizontal(float, float, DWORD = D3DCOLOR_XRGB(255, 0, 0), float = 2);
+	void drawLineHorizontal(int, int, DWORD = D3DCOLOR_XRGB(255, 0, 0), int = 2);
 
-	void drawLineVertical(float, float, DWORD = D3DCOLOR_XRGB(255, 0, 0), float = 2);
+	void drawLineVertical(int, int, DWORD = D3DCOLOR_XRGB(255, 0, 0), int = 2);
 
 };
